First steps to create reverse configs
The use case for this is when you have some songs that are already tagged more or less correctly, and you want to preserve those tags by putting them in a config file.
The idea is you make a quick config file that just says "reverse: full" or "reverse: minimal" and the entire file will be replaced with a normal config that includes all the existing metadata. "full" explicitly outputs everything, where "minimal" won't include fields if an above config is already setting it to that value. Either way, the songs will be untouched after the program runs.
Main challenge right now is serializing a MusicItemConfig. Currently, it's only deserializable, along with tons of YAML data structures like song selectors, strategies, and so on. Ideally, the outputted config won't be redundant, like if all the songs have the same album, it should go in the "songs" strategy, not duplicated in a bunch of individual "set" strategies.
In theory, we don't need every possible implementation of the YAML data structure interfaces to be serializable, only the ones we create ourselves while serializing the config.